requirements.txt
telethon
redis
pyyaml
